Move Nav Bar to components dir
